requirements.txt
supabase
requests
pywhatkit
python-dotenv